+++ integration_tests/tests/conftest.py
@@ -11,7 +11,6 @@ from time import sleep
 
 import pytest
 import os
-import requests
 import zmq
 
 import flowmachine
@@ -137,6 +136,30 @@ def access_token_builder():
         return make_token("test", secret, timedelta(seconds=90), claims)
 
     return token_maker
+
+
+from .utils import query_kinds, permissions_types, aggregation_types
+
+
+@pytest.fixture
+def universal_access_token(access_token_builder):
+    all_claims = {
+        query_kind: {
+            "permissions": permissions_types,
+            "spatial_aggregation": aggregation_types,
+        }
+        for query_kind in query_kinds
+    }
+    all_claims.update(
+        {
+            "geography": {
+                "permissions": permissions_types,
+                "spatial_aggregation": aggregation_types,
+            },
+            "available_dates": {"permissions": {"get_result": True}},
+        }
+    )
+    return access_token_builder(all_claims)
 
 
 @pytest.fixture(scope="session")